# server/openapi.yaml
openapi: 3.0.3
info:
  title: LaunchQueue API
  version: 1.0.0
  description: >
    Production REST API documentation for LaunchQueue — the viral waitlist
    and referral marketing engine for modern SaaS launches.
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.launchqueue.dev
    description: Production API server
paths:
...
  /api/telemetry/vitals:
    post:
      summary: Record a bounded anonymous Web Vitals batch
      description: Requires the configured frontend Origin. Maximum three measurements and a 2 KB text/plain beacon; fixed route templates only. No authentication or identity fields. Rate limited to 30 requests per minute per IP.
      tags: [Telemetry]
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              example: '{"metrics":[{"name":"LCP","route":"/","value":1200}]}'
      responses:
        '204': { description: Accepted }
        '400': { description: Invalid measurements }
        '403': { description: Untrusted origin }
        '413': { description: Payload too large }
        '429': { description: Rate limit exceeded }
...
  /api/waitlists:
    post:
      summary: Create a new waitlist campaign
      tags: [Waitlists]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWaitlistRequest'
      responses:
        '201':
          description: Waitlist campaign created
          content:
            application/json:
              schema:
                type: object
                properties:
                  waitlist:
                    $ref: '#/components/schemas/Waitlist'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Unauthorized

    get:
      summary: List all waitlist campaigns owned by authenticated founder
      tags: [Waitlists]
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: List of waitlists with subscriber counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  waitlists:
                    type: array
                    items:
                      $ref: '#/components/schemas/WaitlistWithCount'
        '401':
          description: Unauthorized
...
  /api/waitlists/{id}:
    get:
      summary: Fetch single waitlist details
      tags: [Waitlists]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Waitlist configuration object
          content:
            application/json:
              schema:
                type: object
                properties:
                  waitlist:
                    $ref: '#/components/schemas/Waitlist'
        '404':
          description: Waitlist not found

    patch:
      summary: Update waitlist branding, milestones, and copy
      tags: [Waitlists]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWaitlistRequest'
      responses:
        '200':
          description: Waitlist updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  waitlist:
                    $ref: '#/components/schemas/Waitlist'
        '400':
          description: Validation error
        '404':
          description: Waitlist not found
...
  /api/payments/webhook:
    post:
      summary: Handle Lemon Squeezy subscription webhook lifecycle events
      tags: [Payments]
      parameters:
        - in: header
          name: x-signature
          required: true
          schema:
            type: string
            description: HMAC SHA-256 signature of the raw request payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                    example: true
        '401':
          description: Invalid HMAC signature

  # ============================================================================
  # Public Waitlist & Referrals (/api/w)
  # ============================================================================
  /api/w/{slug}:
    get:
      summary: Retrieve public waitlist campaign metadata and branding
      tags: [Public Waitlist]
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public campaign details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWaitlistInfo'
        '404':
          description: Waitlist not found
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateWaitlistRequest:
      allOf:
        - $ref: '#/components/schemas/UpdateWaitlistRequest'
        - type: object
          required: [name]
    UpdateWaitlistRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        thankYouMessage:
          type: string
        paused:
          type: boolean
        heroHeadline:
          type: string
        heroSubheadline:
          type: string
        heroImageUrl:
          type: string
        accentColor:
          type: string
        ctaText:
          type: string
        features:
          type: array
          items:
            $ref: '#/components/schemas/FeatureItem'
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/MilestoneItem'
        pageDesign:
          $ref: '#/components/schemas/PageDesign'
    FeatureItem:
      type: object
      required: [title]
      properties:
        icon:
          type: string
          example: "⚡"
        title:
          type: string
          example: Instant Transfers
        description:
          type: string
          example: Zero settlement delay
    MilestoneItem:
      type: object
      required: [referrals, reward]
      properties:
        referrals:
          type: integer
          example: 3
        reward:
          type: string
          example: Priority beta access
    Waitlist:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        paused:
          type: boolean
        features:
          type: array
          items:
            $ref: '#/components/schemas/FeatureItem'
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/MilestoneItem'
        createdAt:
          type: string
          format: date-time
        pageDesign:
          $ref: '#/components/schemas/PageDesign'
    WaitlistWithCount:
      allOf:
        - $ref: '#/components/schemas/Waitlist'
        - type: object
          properties:
            signupCount:
              type: integer
              example: 142
    PublicWaitlistInfo:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        slug:
          type: string
        paused:
          type: boolean
        totalSignups:
          type: integer
        heroHeadline:
          type: string
        heroSubheadline:
          type: string
        heroImageUrl:
          type: string
        accentColor:
          type: string
        ctaText:
          type: string
        features:
          type: array
          items:
            $ref: '#/components/schemas/FeatureItem'
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/MilestoneItem'
        pageDesign:
          $ref: '#/components/schemas/PageDesign'
...
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Please enter a valid email address
        details:
          type: object
          additionalProperties:
            type: string
          example:
            email: Please enter a valid email address
            password: Password must be at least 6 characters
...
    PageDesign:
      type: object
      properties:
        layout:
          type: string
          enum:
            - centered
            - split
            - editorial
        backgroundColor:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
        surfaceColor:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
        fontFamily:
          type: string
          enum:
            - sans
            - serif
            - mono
        cornerStyle:
          type: string
          enum:
            - sharp
            - rounded
            - pill
        backgroundStyle:
          type: string
          enum:
            - solid
            - gradient
            - grid
        eyebrow:
          type: string
          maxLength: 80
        signupHeading:
          type: string
          maxLength: 100
        featureHeading:
          type: string
          maxLength: 120
        rewardHeading:
          type: string
          maxLength: 120
        logoUrl:
          anyOf:
            -
              type: string
              enum:
                - ""
            -
              type: string
              maxLength: 2000
              format: uri
        sectionOrder:
          maxItems: 6
          type: array
          items:
            type: string
            enum:
              - features
              - story
              - steps
              - faq
              - rewards
              - leaderboard
        story:
          type: object
          properties:
            title:
              type: string
              maxLength: 120
            body:
              type: string
              maxLength: 1500
          required:
            - title
            - body
          additionalProperties: false
        steps:
          maxItems: 6
          type: array
          items:
            type: object
            properties:
              title:
                type: string
                minLength: 1
                maxLength: 120
              description:
                type: string
                maxLength: 600
            required:
              - title
              - description
            additionalProperties: false
        faq:
          maxItems: 6
          type: array
          items:
            type: object
            properties:
              question:
                type: string
                minLength: 1
                maxLength: 160
              answer:
                type: string
                maxLength: 600
            required:
              - question
              - answer
            additionalProperties: false
      required:
        - layout
        - backgroundColor
        - surfaceColor
        - fontFamily
        - cornerStyle
        - backgroundStyle
        - eyebrow
        - signupHeading
        - featureHeading
        - rewardHeading
        - logoUrl
        - sectionOrder
        - story
        - steps
        - faq
      additionalProperties: false
...